Move to proper app arch
