+++ lib/data/models/breath_step.dart
@@ -1,0 +1,1 @@
+typedef BreathStep = ({double breathTo, Duration duration});
