+++ pubspec.yaml
@@ -9,6 +9,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_riverpod: ^2.6.1
 
 dev_dependencies:
   flutter_lints: ^4.0.0
